requirements.txt
anthropic>=0.40.0
openai>=1.40.0
requests>=2.31.0
feedparser>=6.0.10
python-dotenv>=1.0.0
markdown>=3.5
